Extract data classes for presentation creation
